+++ sdks/typescript/api/resources/addresses/client/Client.ts
@@ -94,7 +94,7 @@ export class Addresses {
             url: core.url.join(
                 (await core.Supplier.get(this._options.baseUrl)) ??
                     (await core.Supplier.get(this._options.environment)) ??
-                    environments.OumlaSdkApiEnvironment.Production,
+                    environments.OumlaSdkApiEnvironment.Mainnet,
                 "api/v1/addresses/organization",
             ),
             method: "GET",
@@ -184,7 +184,7 @@ export class Addresses {
             url: core.url.join(
                 (await core.Supplier.get(this._options.baseUrl)) ??
                     (await core.Supplier.get(this._options.environment)) ??
-                    environments.OumlaSdkApiEnvironment.Production,
+                    environments.OumlaSdkApiEnvironment.Mainnet,
                 `api/v1/addresses/${encodeURIComponent(reference)}`,
             ),
             method: "GET",
@@ -262,7 +262,7 @@ export class Addresses {
             url: core.url.join(
                 (await core.Supplier.get(this._options.baseUrl)) ??
                     (await core.Supplier.get(this._options.environment)) ??
-                    environments.OumlaSdkApiEnvironment.Production,
+                    environments.OumlaSdkApiEnvironment.Mainnet,
                 "api/v1/addresses/generate",
             ),
             method: "POST",
@@ -343,7 +343,7 @@ export class Addresses {
             url: core.url.join(
                 (await core.Supplier.get(this._options.baseUrl)) ??
                     (await core.Supplier.get(this._options.environment)) ??
-                    environments.OumlaSdkApiEnvironment.Production,
+                    environments.OumlaSdkApiEnvironment.Mainnet,
                 "api/v2/addresses/generate",
             ),
             method: "POST",