+++ sdks/typescript/api/resources/assets/client/Client.ts
@@ -116,7 +116,7 @@ export class Assets {
             url: core.url.join(
                 (await core.Supplier.get(this._options.baseUrl)) ??
                     (await core.Supplier.get(this._options.environment)) ??
-                    environments.OumlaSdkApiEnvironment.Production,
+                    environments.OumlaSdkApiEnvironment.Mainnet,
                 "api/v1/assets",
             ),
             method: "GET",