+++ sdks/typescript/api/resources/withdrawals/client/Client.ts
@@ -87,7 +87,7 @@ export class Withdrawals {
             url: core.url.join(
                 (await core.Supplier.get(this._options.baseUrl)) ??
                     (await core.Supplier.get(this._options.environment)) ??
-                    environments.OumlaSdkApiEnvironment.Production,
+                    environments.OumlaSdkApiEnvironment.Mainnet,
                 "api/v1/withdraw",
             ),
             method: "POST",