+++ sdks/typescript/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.3] - 2025-12-23
+
+### Changed
+- Renamed environment from `Production` to `Mainnet` for better clarity
+  - Use `OumlaSdkApiEnvironment.Mainnet` instead of `OumlaSdkApiEnvironment.Production`
+
+---
+
 ## [1.2.2] - 2025-12-15
 
 ### Fixed
@@ -19,7 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
-## [1.2.0] - 2024-12-12
+## [1.2.0] - 2025-12-12
 
 ### Added
 
@@ -52,7 +60,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
-## [1.0.0] - 2024-11-13
+## [1.0.0] - 2025-11-13
 
 ### Added
 - Initial release of Oumla TypeScript SDK